NOJIRA Rename 'app' module to 'snakeskin'
