+++ .bowerrc
@@ -1,3 +1,3 @@
 {
-   "directory" : "app/static/lib"
+   "directory" : "snakeskin/static/lib"
 }
