+++ run.py
@@ -1,4 +1,4 @@
-from app import app
+from snakeskin import app
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', port=5000, debug=True)
