+++ snakeskin/models/base.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from app import db
+from snakeskin import db
 
 
 """
