+++ snakeskin/models/opt_out.py
@@ -1,5 +1,5 @@
-from app import db
-from app.models.base import Base
+from snakeskin import db
+from snakeskin.models.base import Base
 
 
 """
