+++ snakeskin/proxies/canvas.py
