+++ snakeskin/static/js/controllers/tenantDetailController.js
