+++ snakeskin/static/js/controllers/userDetailController.js
