+++ snakeskin/static/js/factories/tenantFactory.js
