+++ snakeskin/static/js/routes.js
