+++ snakeskin/static/templates/tenantDetail.html
