+++ snakeskin/static/templates/userDetail.html
